Add files via upload
This project focuses on developing a predictive model to assist Carvana in purchasing used cars at auctions while minimizing the risk of acquiring faulty vehicles. The dataset, sourced from Kaggle, includes 32 independent variables and 72,983 instances, with a binary outcome indicating whether a purchase is good or bad. The project involves data preprocessing, feature engineering, and the application of machine learning algorithms such as k-NN, Naive Bayes, and Random Forest. The goal is to enhance Carvana's decision-making process, ensuring the acquisition of high-quality vehicles for resale.